refactor folder structure to align with go package
