+++ Dockerfile
@@ -2,5 +2,5 @@ FROM gcr.io/distroless/static:latest
 LABEL maintainers="MayaData Authors"
 LABEL description="Storage Provisioner"
 
-COPY ./storage-provisioner storage-provisioner
-ENTRYPOINT ["/storage-provisioner"]
+COPY ./d-provisioner d-provisioner
+ENTRYPOINT ["/d-provisioner"]
