+++ hack/update-codegen.sh
@@ -27,7 +27,7 @@ CODEGEN_PKG=${CODEGEN_PKG:-$(cd "${SCRIPT_ROOT}"; ls -d -1 ./vendor/k8s.io/code-
 #                  k8s.io/kubernetes. The output-base is needed for the generators to output into the vendor dir
 #                  instead of the $GOPATH directly. For normal projects this can be dropped.
 bash "${CODEGEN_PKG}"/generate-groups.sh "deepcopy,client,informer,lister" \
-  github.com/AmitKumarDas/storage-provisioner/pkg/client/generated \
+  github.com/AmitKumarDas/storage-provisioner/client/generated \
   github.com/AmitKumarDas/storage-provisioner/pkg/apis \
   ddp:v1alpha1 \
   --go-header-file "${SCRIPT_ROOT}"/hack/custom-boilerplate.go.txt
